+++ src/main/java/roomescape/auth/AdminAuthHandlerInterceptor.java
@@ -5,8 +5,8 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.domain.Member;
-import roomescape.domain.Role;
+import roomescape.domain.member.Member;
+import roomescape.domain.member.Role;
 import roomescape.exception.AuthenticationException;
 import roomescape.service.auth.AuthService;
 import roomescape.service.auth.TokenProvider;
