+++ src/main/java/roomescape/controller/api/AuthApiController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthenticatedMember;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 import roomescape.service.auth.AuthService;
 import roomescape.service.dto.request.LoginRequest;
 import roomescape.service.dto.response.member.MemberIdAndNameResponse;
