+++ src/main/java/roomescape/controller/api/admin/AdminMemberApiController.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 import roomescape.service.dto.response.member.MemberIdAndNameResponses;
 import roomescape.service.member.MemberService;
