+++ src/main/java/roomescape/domain/member/Role.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.member;
 
 public enum Role {
