+++ src/main/java/roomescape/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package roomescape.repository;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberEmail;
 import roomescape.domain.member.MemberPassword;
