+++ src/main/java/roomescape/service/member/MemberService.java
@@ -2,7 +2,7 @@ package roomescape.service.member;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 import roomescape.repository.MemberRepository;
 
 @Service
